+++ gems/aws-sdk-core/lib/aws-sdk-sts/presigner.rb
@@ -53,7 +53,10 @@ module Aws
         )
 
         url = Aws::Partitions::EndpointProvider.resolve(
-          req.context.config.region, 'sts', 'regional'
+          req.context.config.region,
+          'sts',
+          req.context.config.sts_regional_endpoints,
+          req.context.config.use_dualstack_endpoint
         )
         url += "/?#{param_list}"