+++ gems/aws-sdk-docdb/lib/aws-sdk-docdb/plugins/cross_region_copying.rb
@@ -42,7 +42,8 @@ module Aws
               credentials_provider: context.config.credentials
             )
             url = Aws::Partitions::EndpointProvider.resolve(
-              signer.region, 'rds'
+              signer.region, 'rds', 'regional',
+              context.config.use_dualstack_endpoint
             )
             url += "?#{param_list}"
             signer.presign_url(