+++ gems/aws-sdk-s3/lib/aws-sdk-s3/plugins/s3_signer.rb
@@ -22,7 +22,9 @@ module Aws
           # S3 removes core's signature_v4 plugin that checks for this
           raise Aws::Errors::MissingRegionError if cfg.region.nil?
 
-          Aws::Partitions::EndpointProvider.signing_region(cfg.region, 's3')
+          Aws::Partitions::EndpointProvider.signing_region(
+            cfg.region, 's3', cfg.use_dualstack_endpoint
+          )
         end
 
         def add_handlers(handlers, cfg)
@@ -154,7 +156,9 @@ module Aws
 
           def custom_endpoint?(resp)
             resolved_suffix = Aws::Partitions::EndpointProvider.dns_suffix_for(
-              resp.context.config.region
+              resp.context.config.region,
+              's3',
+              resp.context[:use_dualstack_endpoint]
             )
             !resp.context.http_request.endpoint.hostname.include?(resolved_suffix)
           end
@@ -225,12 +229,16 @@ module Aws
           # Otherwise it will retry with the ARN as the bucket name.
           def new_hostname(context, region)
             uri = URI.parse(
-              Aws::Partitions::EndpointProvider.resolve(region, 's3')
+              Aws::Partitions::EndpointProvider.resolve(
+                region, 's3', 'regional', context[:use_dualstack_endpoint]
+              )
             )
 
             if (arn = context.metadata[:s3_arn])
               # Retry with the response region and not the ARN resolved one
-              ARN.resolve_url!(uri, arn[:arn], region).host
+              ARN.resolve_url!(
+                uri, arn[:arn], region, arn[:fips], arn[:dualstack]
+              ).host
             else
               "#{context.params[:bucket]}.#{uri.host}"
             end